+++ integration/js/MeetingLeaveVideoTest.js
@@ -14,34 +14,63 @@ class MeetingLeaveVideoTest extends SdkBaseTest {
   async runIntegrationTest() {
     this.url = this.baseUrl + '?m=' + uuidv4();
     this.numberOfParticipant = 2;
+
+    if (this.numberOfSessions(this.capabilities.browserName) > 1) {
+      await this.runTestOnMultipleSessions()
+    } else {
+      await this.runTestOnSingleSessions()
+    }
+  }
+
+  async runTestOnMultipleSessions() {
     const test_attendee_id = uuidv4();
     const monitor_attendee_id = uuidv4();
 
-    const test_window = await Window.existing(this.driver, "TEST");
-    const monitor_window = await Window.openNew(this.driver, "MONITOR");
+    const testSession = this.seleniumSessions[0];
+    testSession.setSessionName("Test");
+    const monitorSession = this.seleniumSessions[1];
+    monitorSession.setSessionName("Monitor");
 
-    await test_window.runCommands(async () => await this.addUserToMeeting(test_attendee_id));
-    await monitor_window.runCommands(async () => await this.addUserToMeeting(monitor_attendee_id));
+    await this.addUserToMeeting(test_attendee_id, testSession);
+    await this.addUserToMeeting(monitor_attendee_id, monitorSession);
+    await RosterCheck.executeStep(this, testSession, 2);
+    await RosterCheck.executeStep(this, monitorSession, 2);
+    await ClickVideoButton.executeStep(this, monitorSession);
+    await RemoteVideoCheck.executeStep(this, testSession, 'VIDEO_ON');
+    await LeaveMeetingStep.executeStep(this, monitorSession);
+    await RemoteVideoCheck.executeStep(this, testSession, 'VIDEO_OFF');
+  }
 
-    await test_window.runCommands(async () => await RosterCheck.executeStep(this, 2));
-    await monitor_window.runCommands(async () => await RosterCheck.executeStep(this, 2));
+  async runTestOnSingleSessions() {
+    const session = this.seleniumSessions[0];
+    const test_attendee_id = uuidv4();
+    const monitor_attendee_id = uuidv4();
 
-    await test_window.runCommands(async () => await ClickVideoButton.executeStep(this));
-    await monitor_window.runCommands(async () => await RemoteVideoCheck.executeStep(this, 'VIDEO_ON'));
+    const test_window = await Window.existing(session.driver, "TEST");
+    const monitor_window = await Window.openNew(session.driver, "MONITOR");
 
+    await test_window.runCommands(async () => await this.addUserToMeeting(test_attendee_id, session));
+    await monitor_window.runCommands(async () => await this.addUserToMeeting(monitor_attendee_id, session));
 
-    await test_window.runCommands(async () => await LeaveMeetingStep.executeStep(this));
-    await monitor_window.runCommands(async () => await RemoteVideoCheck.executeStep(this, 'VIDEO_OFF'));
+    await test_window.runCommands(async () => await RosterCheck.executeStep(this, session, 2));
+    await monitor_window.runCommands(async () => await RosterCheck.executeStep(this, session, 2));
+
+    await test_window.runCommands(async () => await ClickVideoButton.executeStep(this, session));
+    await monitor_window.runCommands(async () => await RemoteVideoCheck.executeStep(this, session, 'VIDEO_ON'));
+
+    await test_window.runCommands(async () => await LeaveMeetingStep.executeStep(this, session));
+    await monitor_window.runCommands(async () => await RemoteVideoCheck.executeStep(this, session, 'VIDEO_OFF'));
 
     await this.waitAllSteps();
   }
 
-  async addUserToMeeting(attendee_id) {
-    await OpenAppStep.executeStep(this);
-    await AuthenticateUserStep.executeStep(this, attendee_id);
-    await UserAuthenticationCheck.executeStep(this);
-    await JoinMeetingStep.executeStep(this);
-    await UserJoinedMeetingCheck.executeStep(this, attendee_id);
+
+  async addUserToMeeting(attendee_id, sessionInfo) {
+    await OpenAppStep.executeStep(this, sessionInfo);
+    await AuthenticateUserStep.executeStep(this, sessionInfo, attendee_id);
+    await UserAuthenticationCheck.executeStep(this, sessionInfo);
+    await JoinMeetingStep.executeStep(this, sessionInfo);
+    await UserJoinedMeetingCheck.executeStep(this, sessionInfo, attendee_id);
   }
 }