+++ integration/js/steps/AuthenticateUserStep.js
@@ -2,13 +2,13 @@ const {KiteTestError, Status} = require('kite-common');
 const AppTestStep = require('../utils/AppTestStep');
 
 class AuthenticateUserStep extends AppTestStep {
-  constructor(kiteBaseTest, attendee_id) {
-    super(kiteBaseTest);
+  constructor(kiteBaseTest, sessionInfo, attendee_id) {
+    super(kiteBaseTest, sessionInfo);
     this.attendee_id = attendee_id;
   }
 
-  static async executeStep(KiteBaseTest, attendee_id) {
-    const step = new AuthenticateUserStep(KiteBaseTest, attendee_id);
+  static async executeStep(KiteBaseTest, sessionInfo, attendee_id) {
+    const step = new AuthenticateUserStep(KiteBaseTest, sessionInfo, attendee_id);
     await step.execute(KiteBaseTest);
   }
 
@@ -21,11 +21,11 @@ class AuthenticateUserStep extends AppTestStep {
   }
 
   async run() {
-    console.log("attendee id: " + this.attendee_id);
+    this.logger("attendee id: " + this.attendee_id);
     await this.page.enterAttendeeName(this.attendee_id);
 
     await this.page.authenticate();
-    console.log("waiting to authenticate");
+    this.logger("waiting to authenticate");
     let authenticationState = await this.page.waitForAuthentication();
     if (authenticationState === 'failed') {
       throw new KiteTestError(Status.FAILED, 'Authentication timeout');