+++ integration/js/steps/WaitForRemoteParticipantsToJoinMeeting.js
@@ -1,12 +1,12 @@
 const AppWaitTestStep = require('../utils/AppWaitTestStep');
 
 class WaitForRemoteParticipantsToJoinMeeting extends AppWaitTestStep {
-  constructor(kiteBaseTest) {
-    super(kiteBaseTest);
+  constructor(kiteBaseTest, sessionInfo) {
+    super(kiteBaseTest, sessionInfo);
   }
 
-  static async executeStep(KiteBaseTest) {
-    const step = new WaitForRemoteParticipantsToJoinMeeting(KiteBaseTest);
+  static async executeStep(KiteBaseTest, sessionInfo) {
+    const step = new WaitForRemoteParticipantsToJoinMeeting(KiteBaseTest, sessionInfo);
     await step.execute(KiteBaseTest);
   }