+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add video help desk tutorial
 - Enable integration tests for travis builds
 - Add ping/pong to screen sharing start code path to ensure socket is viable
+- Enable integration tests for safari 12
 
 ### Changed
 - Enforce SDP to have candidates for FinishGatheringICECandidateTask to resolve